requirements.txt
python-pptx
pillow